Use Yaml instead of JSON
JSON return unicode strings under Python2, even if this is a better
behavior it would be a breaking change compare to the previous version
of the Agent not using this lib to embed Python. Yaml returns bytes
under Python2 and Unicode under Python3. Since JSON is a subset of Yaml
we use 'yaml' here.